[VAS-738] Moved files in right test folder
