+++ src/pages/commisionalBundles/addEditCommissionBundle/components/drawer/__tests__/BundleTaxonomiesDrawer.test.tsx
@@ -1,13 +1,12 @@
 import { cleanup, render, waitFor, screen, fireEvent } from '@testing-library/react';
 import React from 'react';
-import { store } from '../../../../redux/store';
-import BundleTaxonomiesDrawer from '../components/drawer/BundleTaxonomiesDrawer';
+import { store } from '../../../../../../redux/store';
+import BundleTaxonomiesDrawer from '../BundleTaxonomiesDrawer';
 import { Provider } from 'react-redux';
 import {
   mockedTaxonomy,
   mockedTaxonomyGroups,
-} from '../../../../services/__mocks__/taxonomyService';
-import userEvent from '@testing-library/user-event';
+} from '../../../../../../services/__mocks__/taxonomyService';
 
 const spyOnGetTaxonomiesGroup = jest.spyOn(
   require('../../../../services/taxonomyService.ts'),
