+++ src/pages/commisionalBundles/detail/components/CommissionBundleDetailConfiguration.tsx
@@ -4,13 +4,13 @@ import { Box } from '@mui/system';
 import { ButtonNaked } from '@pagopa/mui-italia';
 import { TitleBox } from '@pagopa/selfcare-common-frontend';
 import { useState } from 'react';
-import { BundleResource, TypeEnum } from '../../../api/generated/portal/BundleResource';
-import { PaddedDrawer } from '../../../components/PaddedDrawer';
+import { BundleResource, TypeEnum } from '../../../../api/generated/portal/BundleResource';
+import { PaddedDrawer } from '../../../../components/PaddedDrawer';
 import {
   formatBooleanValueToYesOrNo,
   formatCurrencyEur,
   formatDateToDDMMYYYY,
-} from '../../../utils/common-utils';
+} from '../../../../utils/common-utils';
 
 const bundleConfigurationFields = {
   col1: [
